current date.js
This program uses the Date object to get the current date, and then formats it based on the desired format. The getDate(), getMonth(), and getFullYear() methods are used to get the day, month, and year of the current date, respectively. The if statements are used to add leading zeros to the month and day if they are less than 10. Finally, the formatted date is printed to the console.